revert: support for doas
The GitHub action I was originally using did not come with sudo
installed and relied on doas which comes in the base OpenBSD install.
I've now switched to a more robust action which includes sudo, among
other things.

This commit reverts support for doas in crt-builder, returning the code
to its much simpler pre-doas state.